requirements.txt
flask
gunicorn
google-generativeai
Pillow
PyMuPDF
werkzeug